Rename : 디렉토리 구조 변경
- 디렉토리 구조 기능 기준으로 변경
