--- be/cocktail/src/main/java/com/BE/cocktail/apiResponse/CocktailRtnGrp.java
@@ -1,14 +1,0 @@
-package com.BE.cocktail.apiResponse;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public enum CocktailRtnGrp {
-
-    Success(201),
-
-    Validation(400);
-
-    private int stautsCode;
-
-}
