--- be/cocktail/src/main/java/com/BE/cocktail/customRecipe/dto/CustomPatchDto.java
@@ -1,25 +1,0 @@
-package com.BE.cocktail.customRecipe.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-public class CustomPatchDto {
-
-    @NotBlank
-    private String imageUrl;
-
-    @NotBlank
-    private String name;
-
-    @NotBlank
-    private String description;
-
-    @NotBlank
-    private String recipe;
-
-    @NotBlank
-    private String ingredient;
-
-}
