+++ be/cocktail/src/main/java/com/BE/cocktail/dto/apiResponse/CocktailRtnGrp.java
@@ -1,0 +1,14 @@
+package com.BE.cocktail.dto.apiResponse;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum CocktailRtnGrp {
+
+    Success(201),
+
+    Validation(400);
+
+    private int stautsCode;
+
+}
