+++ be/cocktail/src/main/java/com/BE/cocktail/dto/regularRecipe/RegularRecipeCreateDto.java
@@ -1,0 +1,28 @@
+package com.BE.cocktail.dto.regularRecipe;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class RegularRecipeCreateDto {
+
+    private String name;
+
+    private String description;
+
+    private String recipe;
+
+    private Integer alcVol;
+
+    private String baseAlc;
+
+    private String ingredient;
+
+    private String imageUrl;
+
+
+}
