+++ be/cocktail/src/main/java/com/BE/cocktail/dto/regularRecipe/RegularRecipeSingleResponseDto.java
@@ -1,0 +1,22 @@
+package com.BE.cocktail.dto.regularRecipe;
+
+import com.BE.cocktail.persistence.domain.regularRecipe.RegularRecipe;
+import com.BE.cocktail.dto.utils.SingleResponse;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+
+//@Getter
+//@AllArgsConstructor(access = AccessLevel.PRIVATE)
+//public class RegularRecipeSingleResponseDto {
+//
+//    private SingleResponse response;
+//
+//    public static RegularRecipeSingleResponseDto of(List<RegularRecipe> regularRecipes) {
+//        RegularRecipeSingleResponseDto response = new RegularRecipeSingleResponseDto(SingleResponse.of(regularRecipes));
+//
+//        return response;
+//    }
+//}
