+++ be/cocktail/src/main/java/com/BE/cocktail/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.BE.cocktail.exception;
 
-import com.BE.cocktail.apiResponse.ApiResponse;
+import com.BE.cocktail.dto.apiResponse.ApiResponse;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
