--- be/cocktail/src/main/java/com/BE/cocktail/regularRecipe/RegularRecipeRepository.java
@@ -1,11 +1,0 @@
-package com.BE.cocktail.regularRecipe;
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
-
-public interface RegularRecipeRepository extends JpaRepository<RegularRecipe, Long> {
-
-    List<RegularRecipe> findAllByName(String name);
-
-}
