+++ be/cocktail/src/test/java/com/BE/cocktail/regularRecipe/RegularRecipeTest.java
@@ -1,11 +1,6 @@
 package com.BE.cocktail.regularRecipe;
 
-import com.BE.cocktail.regularRecipe.dto.RegularRecipeCreateDto;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class RegularRecipeTest {
